refactor: move funcao de finalize pra store
